[LIB-164] Rename sensortreeitem to sensorsettreeitem
