+++ MNE/disp3D/engine/model/data3Dtreemodel.cpp
@@ -40,7 +40,7 @@
 
 #include "data3Dtreemodel.h"
 #include "items/bem/bemtreeitem.h"
-#include "items/sensorspace/sensortreeitem.h"
+#include "items/sensorspace/sensorsettreeitem.h"
 #include "items/subject/subjecttreeitem.h"
 #include "items/freesurfer/fssurfacetreeitem.h"
 #include "items/sourcespace/sourcespacetreeitem.h"
@@ -358,12 +358,12 @@ BemTreeItem* Data3DTreeModel::addBemData(const QString& subject, const QString& 
 
 //*************************************************************************************************************
 
-SensorTreeItem* Data3DTreeModel::addMegSensorData(const QString& subject,
-                                                   const QString& sSensorSetName,
-                                                   const MNELIB::MNEBem& tSensor,
-                                                   const QList<FIFFLIB::FiffChInfo>& lChInfo)
+SensorSetTreeItem* Data3DTreeModel::addMegSensorData(const QString& subject,
+                                                       const QString& sSensorSetName,
+                                                       const MNELIB::MNEBem& tSensor,
+                                                       const QList<FIFFLIB::FiffChInfo>& lChInfo)
 {
-    SensorTreeItem* pReturnItem = Q_NULLPTR;
+    SensorSetTreeItem* pReturnItem = Q_NULLPTR;
 
     //Handle subject item
     SubjectTreeItem* pSubjectItem = addSubject(subject);
@@ -371,11 +371,11 @@ SensorTreeItem* Data3DTreeModel::addMegSensorData(const QString& subject,
     //Find already existing surface items and add the new data to the first search result
     QList<QStandardItem*> itemList = pSubjectItem->findChildren(sSensorSetName);
 
-    if(!itemList.isEmpty() && (itemList.first()->type() == Data3DTreeModelItemTypes::SensorItem)) {
-        pReturnItem = dynamic_cast<SensorTreeItem*>(itemList.first());
+    if(!itemList.isEmpty() && (itemList.first()->type() == Data3DTreeModelItemTypes::SensorSetItem)) {
+        pReturnItem = dynamic_cast<SensorSetTreeItem*>(itemList.first());
         pReturnItem->addData(tSensor, lChInfo, m_pModelEntity);
     } else {
-        pReturnItem = new SensorTreeItem(Data3DTreeModelItemTypes::SensorItem, sSensorSetName);
+        pReturnItem = new SensorSetTreeItem(Data3DTreeModelItemTypes::SensorSetItem, sSensorSetName);
         addItemWithDescription(pSubjectItem, pReturnItem);
         pReturnItem->addData(tSensor, lChInfo, m_pModelEntity);
     }
