+++ MNE/disp3D/engine/model/items/common/types.h
@@ -96,7 +96,7 @@ namespace Data3DTreeModelItemTypes
                     DigitizerItem = QStandardItem::UserType + 12,
                     ECDDataItem = QStandardItem::UserType + 13,
                     MriItem = QStandardItem::UserType + 14,
-                    SensorItem = QStandardItem::UserType + 15,
+                    SensorSetItem = QStandardItem::UserType + 15,
                     SensorSurfaceItem = QStandardItem::UserType + 16};
 }
